'Crear_Seccion'

+++ app/Gondola.php
@@ -4,12 +4,16 @@ namespace App;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Eloquent\Model;
 use App\Producto;
+use App\Empleado_Tiene_Gondola;
+use App\Gondola_Tiene_Seccion;
+
 class Gondola extends Model
 {
 
     protected $connection = 'retail';
     protected $table = 'gondolas';
     protected $primaryKey='Codigo';
+    public $timestamps = false;
     const CREATED_AT='FECALTAS';
     const UPDATED_AT='FECMODIF';
 
@@ -45,11 +49,19 @@ class Gondola extends Model
     }
         public static function filtrar_gondola($datos)
     {
+        
         /*  --------------------------------------------------------------------------------- */
         
         $user = auth()->user();
         // OBTENER TODOS LOS DATOS DEL TALLE
-        $gondolas = Gondola::select(DB::raw('ID AS CODIGO, DESCRIPCION'))->where('ID_SUCURSAL','=',$user->id_sucursal)->Where('ID','=',$datos['id'])->get()->toArray();
+
+        $gondolas = Gondola::leftjoin('gondola_tiene_seccion', 'gondola_tiene_seccion.ID_GONDOLA', '=', 'GONDOLAS.ID')
+        ->select(DB::raw('gondolas.ID as CODIGO,DESCRIPCION, gondola_tiene_seccion.ID_SECCION AS ID_SECCION'))
+        ->where('gondolas.ID_SUCURSAL','=',$user->id_sucursal)
+        ->Where('gondolas.ID','=',$datos['id'])
+        ->get()
+        ->toArray();
+
         if(count($gondolas)<=0){
            return ["response"=>false];
         }
@@ -84,6 +96,7 @@ class Gondola extends Model
         $columns = array( 
                             0 => 'codigo', 
                             1 => 'descripcion',
+                            2=> 'seccion'
                          
                         );
         
@@ -113,12 +126,14 @@ class Gondola extends Model
 
             //  CARGAR TODOS LOS PRODUCTOS ENCONTRADOS 
 
-            $posts = Gondola::select(DB::raw('ID AS CODIGO,DESCRIPCION'))
-                         ->where('id_sucursal','=',$user->id_sucursal)
-                         ->offset($start)
-                         ->limit($limit)
-                         ->orderBy($order,$dir)
-                         ->get();
+            $posts = Gondola::select(DB::raw('GONDOLAS.ID AS CODIGO,GONDOLAS.DESCRIPCION, IFNULL(SECCIONES.DESCRIPCION,"NO POSEE") AS SECCION'))
+                ->leftjoin('gondola_tiene_seccion', 'gondola_tiene_seccion.ID_GONDOLA', '=', 'GONDOLAS.ID')
+                ->leftjoin('secciones', 'secciones.ID', '=', 'gondola_tiene_seccion.ID_SECCION')
+                ->where('Gondolas.id_sucursal','=',$user->id_sucursal)
+                ->offset($start)
+                ->limit($limit)
+                ->orderBy($order,$dir)
+                ->get();
 
             /*  ************************************************************ */
 
@@ -134,11 +149,14 @@ class Gondola extends Model
 
             // CARGAR LOS PRODUCTOS FILTRADOS EN DATATABLE
 
-            $posts =Gondola::select(DB::raw('ID AS CODIGO,DESCRIPCION'))
-                            ->where('id_sucursal','=',$user->id_sucursal)
+            $posts =Gondola::select(DB::raw('GONDOLAS.ID AS CODIGO,GONDOLAS.DESCRIPCION, IFNULL(SECCIONES.DESCRIPCION,"NO POSEE") AS SECCION'))
+                ->leftjoin('gondola_tiene_seccion', 'gondola_tiene_seccion.ID_GONDOLA', '=', 'GONDOLAS.ID')
+                ->leftjoin('secciones', 'secciones.ID', '=', 'gondola_tiene_seccion.ID_SECCION')
+                            ->where('GONDOLAS.id_sucursal','=',$user->id_sucursal)
                             ->where(function ($query) use ($search) {
-                                $query->where('ID','LIKE',"%{$search}%")
-                                      ->orWhere('DESCRIPCION', 'LIKE',"%{$search}%");
+                                $query->where('GONDOLAS.ID','LIKE',"%{$search}%")
+                                      ->orWhere('GONDOLAS.DESCRIPCION', 'LIKE',"%{$search}%")
+                                      ->orWhere('SECCIONES.DESCRIPCION', 'LIKE',"%{$search}%");
                             })
                             ->offset($start)
                             ->limit($limit)
@@ -149,10 +167,13 @@ class Gondola extends Model
 
             // CARGAR LA CANTIDAD DE PRODUCTOS FILTRADOS 
 
-            $totalFiltered = Gondola::where('id_sucursal','=',$user->id_sucursal)
-                          ->where(function ($query) use ($search) {
-                                $query->where('ID','LIKE',"%{$search}%")
-                                      ->orWhere('DESCRIPCION', 'LIKE',"%{$search}%");
+            $totalFiltered = Gondola::leftjoin('gondola_tiene_seccion', 'gondola_tiene_seccion.ID_GONDOLA', '=', 'GONDOLAS.ID')
+                ->leftjoin('secciones', 'secciones.ID', '=', 'gondola_tiene_seccion.ID_SECCION')
+                            ->where('GONDOLAS.id_sucursal','=',$user->id_sucursal)
+                            ->where(function ($query) use ($search) {
+                                $query->where('GONDOLAS.ID','LIKE',"%{$search}%")
+                                      ->orWhere('GONDOLAS.DESCRIPCION', 'LIKE',"%{$search}%")
+                                      ->orWhere('SECCIONES.DESCRIPCION', 'LIKE',"%{$search}%");
                             })
                              ->count();
 
@@ -177,7 +198,7 @@ class Gondola extends Model
 
                 $nestedData['CODIGO'] = $post->CODIGO;
                 $nestedData['DESCRIPCION'] = $post->DESCRIPCION;
-
+                $nestedData['SECCION'] = $post->SECCION;
 
                 $data[] = $nestedData;
 
@@ -229,49 +250,52 @@ blob:https://web.whatsapp.com/3c60c7d0-5c70-40fc-93b4-53017c2e03ef
         /*  --------------------------------------------------------------------------------- */
 
     }
-         public static function gondola_guardar($datos)
-    {
+
+
+    public static function gondola_guardar($datos){
 
         $user = auth()->user();
         $dia = date("Y-m-d H:i:s");
         $hora = date("H:i:s");
         $codigo_gondola=$datos['data']['Codigo'];
+        $seccion_id = $datos['data']['SeccionGuardar'];
             //var_dump($datos['data']['Marcados']);
         try { 
+            DB::connection('retail')->beginTransaction();
         /*  --------------------------------------------------------------------------------- */
-         if($datos['data']['Existe']=== false){
-         $gondolas =Gondola::insertGetId(
-         ['DESCRIPCION'=> $datos['data']['Descripcion'], 'FK_USER_CR'=>$user->id,'FECALTAS'=>$dia,'ID_SUCURSAL'=>$user->id_sucursal]);
-     
-
-  
-
-
-        }else{
-             //GONDOLAS UPDATE
-            $gondolas=Gondola::where('ID', $codigo_gondola)
-            ->update(['DESCRIPCION'=> $datos['data']['Descripcion'], 'FK_USER_MD'=>$user->id,'FECMODIF'=>$dia]);
-
-
- 
-           }
-
-
-            return ["response"=>true];
-        
-      
-
+            if($datos['data']['Existe']=== false){
+                $gondolas =Gondola::insertGetId([
+                    'DESCRIPCION'=> $datos['data']['Descripcion'], 
+                    'FK_USER_CR'=>$user->id,
+                    'FECALTAS'=>$dia,
+                    'ID_SUCURSAL'=>$user->id_sucursal]);
+                DB::connection('retail')->commit();
+                if($seccion_id!='null' || $seccion_id!=''){
+                    $gondola_id=$gondolas;
+                    Gondola_Tiene_Seccion::asignar_seccion($gondola_id, $seccion_id,$user->id_sucursal);
+                }
+                return ["response"=>true];
+            }else{
+                 //GONDOLAS UPDATE
+                $gondolas=Gondola::where('ID', '=' ,$codigo_gondola)
+                ->update([
+                    'DESCRIPCION'=> $datos['data']['Descripcion'], 
+                    'FK_USER_MD'=>$user->id,
+                    'FECMODIF'=>$dia]);
+                Gondola_Tiene_Seccion::modificar_seccion($seccion_id, $codigo_gondola, $user->id_sucursal);
+                DB::connection('retail')->commit();
+                return ["response"=>true];
+            }
+           
         /*  --------------------------------------------------------------------------------- */
-        } catch(Exception $ex){ 
-
- 
-        if($ex->errorInfo[1]===1062){
-      return ["response"=>false,'statusText'=>'Esta Codigo de gondola ya fue registrada!!'];
-       }else{
-      return ["response"=>false,'statusText'=>$ex->getMessage()];
-      }
-  
-      }
+        }catch(Exception $ex){
+            DB::connection('retail')->rollBack(); 
+            if($ex->errorInfo[1]===1062){
+                return ["response"=>false,'statusText'=>'Esta Codigo de gondola ya fue registrada!!'];
+            }else{
+                return ["response"=>false,'statusText'=>$ex->getMessage()];
+            }
+        }
 
 
     }
@@ -284,22 +308,27 @@ blob:https://web.whatsapp.com/3c60c7d0-5c70-40fc-93b4-53017c2e03ef
         /*  --------------------------------------------------------------------------------- */
 
         if($datos['data']['Existe']=== true){
-        
-         $producto=Producto::select('CODIGO')
-         ->leftjoin('GONDOLA_TIENE_PRODUCTOS','GONDOLA_TIENE_PRODUCTOS.GONDOLA_COD_PROD','=','productos.CODIGO')
-         ->Where('GONDOLA_TIENE_PRODUCTOS.ID_GONDOLA','=',$codigo_gondola)->limit(1)->get()->toArray();
-         if(count($producto)>0){
-            return ["response"=>false];
-         }else{
+            $producto=Producto::select('Productos.CODIGO')
+                ->leftjoin('GONDOLA_TIENE_PRODUCTOS','GONDOLA_TIENE_PRODUCTOS.GONDOLA_COD_PROD','=','productos.CODIGO')
+                ->Where('GONDOLA_TIENE_PRODUCTOS.ID_GONDOLA','=',$codigo_gondola)->limit(1)->get()->toArray();
+            if(count($producto)>0){
+                return ["response"=>false, "statusText"=>"Este codigo de gondola no existe o tiene productos asignados"];
+            }
+             $empleado=Empleado_Tiene_Gondola::select('empleado_tiene_gondola.FK_GONDOLA')
+                ->leftjoin('Gondolas','Gondolas.ID','=','empleado_tiene_gondola.FK_GONDOLA')
+                ->Where('empleado_tiene_gondola.FK_GONDOLA','=',$codigo_gondola)->limit(1)->get()->toArray();
+            if(count($empleado)>0){
+                return ["response"=>false, "statusText"=>"Este codigo de gondola no existe o tiene empleados asignados"];
+            }
 
-
-        $gondola=Gondola::where('ID', $codigo_gondola)->delete();
-                    }
+            gondola_tiene_seccion::where('ID_GONDOLA','=',$codigo_gondola)->delete();  
+            $gondola=Gondola::where('ID', $codigo_gondola)->delete();
+            return ["response"=>true];
 
         }else{
-        return ["response"=>false];
+            return ["response"=>false, "statusText"=>"Este codigo gondola no existe"];
         }
-  return ["response"=>true];
+        
 
         /*  --------------------------------------------------------------------------------- */
